Move src contents into react-app
