+++ packages/react-app/src/App.js
@@ -1,73 +1,139 @@
 import React from "react";
-import { Contract } from "@ethersproject/contracts";
-import { getDefaultProvider } from "@ethersproject/providers";
-import { useQuery } from "@apollo/react-hooks";
+import { Button, Slider, Switch } from "antd";
+import "./App.css";
+import "antd/dist/antd.css";
+import {
+  Layout,
+  Menu,
+  Breadcrumb,
+  Form,
+  Checkbox,
+  Input,
+  Row,
+  Col,
+  Tooltip,
+} from "antd";
+import { PageHeader } from "antd";
+import TradingViewWidget, { Themes } from "react-tradingview-widget";
+import { AdvancedChart } from "react-tradingview-widget";
+import { Dropdown, message } from "antd";
+import { DownOutlined } from "@ant-design/icons";
+import { Descriptions } from "antd";
+import { InfoCircleOutlined, UserOutlined } from "@ant-design/icons";
+import { DownloadOutlined } from "@ant-design/icons";
 
-import { Body, Button, Header, Image, Link } from "./components";
-import logo from "./ethereumLogo.png";
-import useWeb3Modal from "./hooks/useWeb3Modal";
+const { Header, Content, Footer } = Layout;
+const onClick = ({ key }) => {
+  message.info(`Click on item ${key}`);
+};
 
-import { addresses, abis } from "@project/contracts";
-import GET_TRANSFERS from "./graphql/subgraph";
+const menu = (
+  <Menu onClick={onClick}>
+    <Menu.Item key="1">1st menu item</Menu.Item>
+    <Menu.Item key="2">2nd menu item</Menu.Item>
+    <Menu.Item key="3">3rd menu item</Menu.Item>
+  </Menu>
+);
 
-async function readOnChainData() {
-  // Should replace with the end-user wallet, e.g. Metamask
-  const defaultProvider = getDefaultProvider();
-  // Create an instance of an ethers.js Contract
-  // Read more about ethers.js on https://docs.ethers.io/v5/api/contract/contract/
-  const ceaErc20 = new Contract(addresses.ceaErc20, abis.erc20, defaultProvider);
-  // A pre-defined address that owns some CEAERC20 tokens
-  const tokenBalance = await ceaErc20.balanceOf("0x3f8CB69d9c0ED01923F11c829BaE4D9a4CB6c82C");
-  console.log({ tokenBalance: tokenBalance.toString() });
-}
-
-function WalletButton({ provider, loadWeb3Modal, logoutOfWeb3Modal }) {
-  return (
-    <Button
-      onClick={() => {
-        if (!provider) {
-          loadWeb3Modal();
-        } else {
-          logoutOfWeb3Modal();
-        }
-      }}
-    >
-      {!provider ? "Connect Wallet" : "Disconnect Wallet"}
-    </Button>
-  );
-}
+const layout = {
+  labelCol: { span: 8 },
+  wrapperCol: { span: 16 },
+};
+const tailLayout = {
+  wrapperCol: { offset: 8, span: 16 },
+};
 
 function App() {
-  const { loading, error, data } = useQuery(GET_TRANSFERS);
-  const [provider, loadWeb3Modal, logoutOfWeb3Modal] = useWeb3Modal();
-
-  React.useEffect(() => {
-    if (!loading && !error && data && data.transfers) {
-      console.log({ transfers: data.transfers });
-    }
-  }, [loading, error, data]);
-
   return (
-    <div>
-      <Header>
-        <WalletButton provider={provider} loadWeb3Modal={loadWeb3Modal} logoutOfWeb3Modal={logoutOfWeb3Modal} />
+    <Layout>
+      <Header style={{ position: "fixed", zIndex: 1, width: "100%" }}>
+        <div className="logo" />
+        <Menu theme="dark" mode="horizontal" defaultSelectedKeys={["2"]}>
+          <Menu.Item key="1">Trade</Menu.Item>
+          <Menu.Item key="2">Staking</Menu.Item>
+          <Menu.Item key="3">Rewards</Menu.Item>
+        </Menu>
       </Header>
-      <Body>
-        <Image src={logo} alt="react-logo" />
-        <p>
-          Edit <code>packages/react-app/src/App.js</code> and save to reload.
-        </p>
-        {/* Remove the "hidden" prop and open the JavaScript console in the browser to see what this function does */}
-        <Button hidden onClick={() => readOnChainData()}>
-          Read On-Chain Balance
-        </Button>
-        <Link href="https://ethereum.org/developers/#getting-started" style={{ marginTop: "8px" }}>
-          Learn Ethereum
-        </Link>
-        <Link href="https://reactjs.org">Learn React</Link>
-        <Link href="https://thegraph.com/docs/quick-start">Learn The Graph</Link>
-      </Body>
-    </div>
+      <Content
+        className="site-layout"
+        style={{ padding: "0 50px", height: 1000, marginTop: 64 }}
+      >
+        <div
+          className="site-layout-background"
+          style={{ padding: 12, minHeight: 380 }}
+        >
+          <Dropdown overlay={menu}>
+            <a
+              className="ant-dropdown-link"
+              onClick={(e) => e.preventDefault()}
+            >
+              EUR/USD <DownOutlined />
+            </a>
+          </Dropdown>
+          <Descriptions title="">
+            <Descriptions.Item label="My Balance">0.00</Descriptions.Item>
+          </Descriptions>
+          <Button type="primary" shape="round" size={10}>
+            BUY/LONG
+          </Button>
+          &nbsp;&nbsp;&nbsp;
+          <Button type="primary" shape="round" size={10}>
+            SELL/SHORT
+          </Button>
+          <Row gutter={[40, 16]}>
+            <Col span={6} />
+            <Col span={6} />
+            <Col span={6} />
+            <Col span={6} />
+
+            <Col span={6} />
+            <Col span={6} />
+            <Col span={6} />
+            <Col span={6} />
+          </Row>
+          <Form>
+            <Form.Item
+              label="Collateral"
+              name="username"
+              rules={[{ message: "Please input your username!" }]}
+            >
+              <Input placeholder="0.00" suffix="USDC" />
+            </Form.Item>
+
+            <Form.Item
+              label="Position"
+              name="password"
+              rules={[{ message: "Please input your password!" }]}
+            >
+              <Input placeholder="0.00" suffix="EURO" />
+            </Form.Item>
+
+            <Form.Item>
+              <Slider
+                defaultValue={3}
+                disabled={false}
+                max={10}
+                marginTop={64}
+              />
+            </Form.Item>
+            <Form.Item>
+              <Button type="primary" htmlType="submit">
+                Submit
+              </Button>
+            </Form.Item>
+          </Form>
+        </div>
+        <TradingViewWidget
+          symbol="FX:EURUSD"
+          theme={Themes.DARK}
+          locale="fr"
+          dateRange={12}
+          width={800}
+          height={400}
+        />
+      </Content>
+      <Footer style={{ textAlign: "center" }}>Hedgit</Footer>
+    </Layout>
   );
 }
