--- packages/react-app/src/graphql/subgraph.js
@@ -1,15 +1,0 @@
-import { gql } from "apollo-boost";
-
-// See more example queries on https://thegraph.com/explorer/subgraph/paulrberg/create-eth-app
-const GET_TRANSFERS = gql`
-  {
-    transfers(first: 10) {
-      id
-      from
-      to
-      value
-    }
-  }
-`;
-
-export default GET_TRANSFERS;
