+++ packages/react-app/src/index.js
@@ -1,19 +1,19 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import ApolloClient from "apollo-boost";
-import { ApolloProvider } from "@apollo/react-hooks";
-import "./index.css";
-import App from "./App";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
 
-// You should replace this url with your own and put it into a .env file
-// See all subgraphs: https://thegraph.com/explorer/
-const client = new ApolloClient({
-  uri: "https://api.thegraph.com/subgraphs/name/paulrberg/create-eth-app",
-});
+import 'antd/dist/antd.css';
+import './index.css';
+import { Layout, Menu, Breadcrumb } from 'antd';
+
+const { Header, Content, Footer } = Layout;
 
 ReactDOM.render(
-  <ApolloProvider client={client}>
-    <App />
-  </ApolloProvider>,
-  document.getElementById("root"),
+  <React.StrictMode>
+      <App />
+
+  </React.StrictMode>,
+  document.getElementById('root')
 );
