--- src/index.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-import 'antd/dist/antd.css';
-import './index.css';
-import { Layout, Menu, Breadcrumb } from 'antd';
-
-const { Header, Content, Footer } = Layout;
-
-ReactDOM.render(
-  <React.StrictMode>
-      <App />
-
-  </React.StrictMode>,
-  document.getElementById('root')
-);
